reactor: Batch remote thread wakeups
This batches remote thread wakeups by changing send_msg() to only set a
bit in a "pending wakeups" bitset and waking up all remote threads once
per reactor loop. The main motivation for this change is to eliminate
the expensive full memory barriers in send_msg() when checking if a
remote thread is sleeping or not.