sphinx: Unify cache line size constants
This introduces a new sphinx::hardware::cache_line_size constant, which
specifies the size of a CPU L1 cache line, and switches call-sites in
server front-end and SPSC queue to use it. We pick L1 cache line size of
64 bytes because it's what x86 has had ever since the Core uarch was
introduced circa 2008 (the precedessor uarch, NetBurst, had a cache line
size of 128 bytes, though).