reactor: Fix partial transmission for TcpSocket::send()
The send() system call often consumes the whole buffer passed to it, but
it is not guaranteed. When the system is under enough load, send() may
consume the buffer partially or reject it altogether with the EAGAIN or
EWOULDBLOCK errors. If that happens, store the buffer internally, and
use EPOLLOUT epoll event to receive notification when the buffer can be
transmitted. However, please note that the common case is still a
non-blocking send() without going through epoll.