sphinxd: Respond to request from remote CPU
Let's respond to a request from the CPU core that serves the request
because that's where the data is. Sending a message back to the core
where request arrived ensures that socket locks are not contended, but
the downside is that we have to copy data from one CPU core cache to
another.